+++ linguee_api/config.py
@@ -13,8 +13,13 @@ class Settings(BaseSettings):
     sentry_dsn: Optional[str] = None
     sentry_environment: str = "development"
 
-    # File cache settings
+    # File and SQLite cache settings
     cache_directory: pathlib.Path = PROJECT_ROOT / ".cache"
+
+    @property
+    def cache_database(self) -> pathlib.Path:
+        """Cache database."""
+        return self.cache_directory / "cache.sqlite3"
 
     class Config:
         env_file = (PROJECT_ROOT / ".env").as_posix()