+++ linguee_api/downloaders/file_cache.py
@@ -1,10 +1,11 @@
 import pathlib
 import urllib.parse
+from typing import Optional
 
-from linguee_api.downloaders.interfaces import IDownloader
+from linguee_api.downloaders.interfaces import ICache, IDownloader
 
 
-class FileCache(IDownloader):
+class FileCache(ICache):
     """File Cache."""
 
     def __init__(self, cache_directory: pathlib.Path, upstream: IDownloader):
@@ -12,12 +13,15 @@ class FileCache(IDownloader):
         self.upstream = upstream
         self.cache_directory.mkdir(parents=True, exist_ok=True)
 
-    async def download(self, url: str) -> str:
-        page_file = self.get_page_file(url)
-        if not page_file.is_file():
-            page = await self.upstream.download(url)
-            page_file.write_text(page, encoding="utf-8")
-        return page_file.read_text(encoding="utf-8")
+    async def get_from_cache(self, url: str) -> Optional[str]:
+        page_file = self._get_page_file(url)
+        if page_file.is_file():
+            return page_file.read_text(encoding="utf-8")
+        return None
 
-    def get_page_file(self, url: str) -> pathlib.Path:
+    async def put_to_cache(self, url: str, page: str) -> None:
+        page_file = self._get_page_file(url)
+        page_file.write_text(page, encoding="utf-8")
+
+    def _get_page_file(self, url: str) -> pathlib.Path:
         return self.cache_directory / urllib.parse.quote(url, safe="")